Implemented deserialize and serialize function for VarLengthMetaAttribute and abstract out a MetaAttribute interface
