+++ src/StorageManager/Metadata/Attribute/FixedLengthMetaAttribute.java
@@ -5,46 +5,47 @@ import Constants.Constant.DataType;
 import Constants.Helper;
 import java.util.Arrays;
 
-public class VarLengthMetaAttribute extends FixedLengthMetaAttribute implements MetaAttribute {
+public class FixedLengthMetaAttribute implements MetaAttribute {
 
-    // length of type (CHAR, VARCHAR);
-    private final int length;
+    private final String name;
+    private final DataType type;
+    private final boolean isPrimaryKey;
 
-    public VarLengthMetaAttribute(String name, Constant.DataType type, Boolean isPrimaryKey,
-        int length) {
-        super(name, type, isPrimaryKey);
-        this.length = length;
+    // byte: length of name, string name, number of length of type, type
+
+    public FixedLengthMetaAttribute(String name, DataType type, boolean isPrimaryKey) {
+        this.name = name;
+        this.type = type;
+        this.isPrimaryKey = isPrimaryKey;
     }
 
     /**
-     * serialize varLengthMetaAttribute into the following form in bytes [isPrimary(boolean),
-     * nameLength(int), name(String), typeLength(int),typeString(String), valueLength(int)];
+     * serialize metaAttribute into the following form in bytes [isPrimary(boolean),
+     * nameLength(int), name(String), typeLength(int),typeString(String)];
      *
-     * @param varLengthMetaAttribute - varLengthMetaAttribute
+     * @param metaAttribute - metaAttribute
      * @return byte arrays
      */
-    public byte[] serialize(VarLengthMetaAttribute varLengthMetaAttribute) {
+    public byte[] serialize(FixedLengthMetaAttribute metaAttribute) {
         byte[] isPrimaryKeyBytes = new byte[]{
-            Helper.convertBooleanToByte(varLengthMetaAttribute.getIsPrimaryKey())};
-        String name = varLengthMetaAttribute.getName();
+            Helper.convertBooleanToByte(metaAttribute.getIsPrimaryKey())};
+        String name = metaAttribute.getName();
         int nameLength = name.length();
         byte[] nameLengthBytes = Helper.convertIntToByteArray(nameLength);
         byte[] nameBytes = Helper.convertStringToByteArrays(name);
 
-        DataType type = varLengthMetaAttribute.getType();
+        DataType type = metaAttribute.getType();
         String typeString = getTypeString(type);
         int typeLength = typeString.length();
         byte[] typeLengthBytes = Helper.convertIntToByteArray(typeLength);
         byte[] typeStringBytes = Helper.convertStringToByteArrays(typeString);
-        byte[] valueLength = Helper.convertIntToByteArray(varLengthMetaAttribute.getLength());
         return Helper.concatenate(isPrimaryKeyBytes, nameLengthBytes, nameBytes, typeLengthBytes,
-            typeStringBytes, valueLength);
+            typeStringBytes);
     }
 
     /**
      * Construct metaAttribute object from the byte array in the following form:
-     * [isPrimary(boolean), nameLength(int), name(String), typeLength(int), typeString(String),
-     * valueLength(int)]
+     * [isPrimary(boolean), nameLength(int), name(String), typeLength(int),typeString(String)]
      *
      * @param bytes - binary form of metaAttribute
      * @return metaAttribute object
@@ -66,27 +67,37 @@ public class VarLengthMetaAttribute extends FixedLengthMetaAttribute implements 
         index += Constant.INTEGER_SIZE + 1;
         String typeString = Helper.convertByteArrayToString(
             Arrays.copyOfRange(bytes, index, index + typeLength + 1));
-        index += typeLength + 1;
         DataType type = switch (typeString) {
-            case Constant.CHAR -> DataType.CHAR;
-            case Constant.VARCHAR -> DataType.VARCHAR;
+            case Constant.INTEGER -> DataType.INTEGER;
+            case Constant.BOOLEAN -> DataType.BOOLEAN;
+            case Constant.DOUBLE -> DataType.DOUBLE;
         };
-        int valueLength = Helper.convertByteArrayToInt(
-            Arrays.copyOfRange(bytes, index, index + Constant.INTEGER_SIZE + 1));
-
-        return new VarLengthMetaAttribute(name, type, isPrimaryKey, valueLength);
+        return new FixedLengthMetaAttribute(name, type, isPrimaryKey);
     }
 
-    @Override
+
     protected String getTypeString(DataType type) {
         return switch (type) {
-            case CHAR -> Constant.CHAR;
-            case VARCHAR -> Constant.VARCHAR;
+            case INTEGER -> Constant.INTEGER;
+            case DOUBLE -> Constant.DOUBLE;
+            case BOOLEAN -> Constant.BOOLEAN;
             default -> throw new IllegalStateException("Unexpected value: " + type);
         };
     }
 
-    public int getLength() {
-        return length;
+    @Override
+    public String getName() {
+        return name;
     }
+
+    @Override
+    public Constant.DataType getType() {
+        return type;
+    }
+
+    @Override
+    public boolean getIsPrimaryKey() {
+        return isPrimaryKey;
+    }
+
 }
